worked on infrastructure part and db migrations

+++ src/Common/SharedKernel/ValueObjects/DimensionsInfo.cs
@@ -8,26 +8,26 @@ public class DimensionsInfo : ValueObject
     /// <summary>
     /// Length in cm
     /// </summary>
-    public decimal Length { get; }
+    public int Length { get; }
 
     /// <summary>
     /// Width in cm
     /// </summary>
-    public decimal Width { get; }
+    public int Width { get; }
 
     /// <summary>
     /// Height in cm
     /// </summary>
-    public decimal Height { get; }
+    public int Height { get; }
 
     /// <summary>
     /// Weight in grams
     /// </summary>
-    public decimal Weight { get; }
+    public int Weight { get; }
 
     public decimal Volume => Length * Width * Height;
 
-    public DimensionsInfo(decimal length, decimal width, decimal height, decimal weight)
+    public DimensionsInfo(int length, int width, int height, int weight)
     {
         Length = length;
         Width = width;
@@ -35,7 +35,7 @@ public class DimensionsInfo : ValueObject
         Weight = weight;
     }
 
-    public static Result<DimensionsInfo> Create(decimal length, decimal width, decimal height, decimal weight)
+    public static Result<DimensionsInfo> Create(int length, int width, int height, int weight)
     {
         if (length <= 0 || width <= 0 || height <= 0)
         {
